requirements/django/requirements.txt
Django==5.1.4
psycopg2-binary==2.9.10
python-decouple==3.8
djangorestframework==3.15.2
pillow==10.3.0
dj-rest-auth==7.0.0
django-allauth==0.54.0
requests==2.31.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
#channels-redis==3.3.0
daphne==4.1.0
#asgiref==3.3.4
